refactor: move discovery fetchers and dataset extractors to cli/commands/discovery.py
